LAB #2/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.2.2
click==8.1.3
colorama==0.4.6
constantly==23.10.4
contourpy==1.0.6
cryptography==45.0.3
cycler==0.11.0
daphne==4.2.0
distlib==0.3.9
Django==5.0.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
drf-yasg==1.21.10
filelock==3.18.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
fonttools==4.38.0
greenlet==2.0.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
platformdirs==4.3.8
prettytable==3.6.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
service-identity==24.2.0
six==1.16.0
SQLAlchemy==1.4.46
sqlparse==0.5.0
tabulate==0.9.0
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.2.0
virtualenv==20.31.2
wcwidth==0.2.5
Werkzeug==2.2.2
zope.interface==7.2
